refactor(cli): remove action-contract re-export shim
Problem: public CLI action contracts live in polylogue.operations.action_contracts, but devtools and tests still imported them through polylogue.cli.action_contracts, a pure compatibility re-export module.

What changed: delete the CLI shim, import the owning operation contract module directly, and regenerate CLI/topology docs for the removed module.

Verification:
- rg -n 'polylogue\.cli\.action_contracts|CLI compatibility imports' . -> no matches
- devtools test tests/unit/cli/test_cli_action_contracts.py tests/unit/cli/test_completion_matrix.py tests/unit/devtools/test_action_contract_report.py -> 110 passed
- devtools verify --quick -> exit_code 0

Ref #1849
