+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 697
+- **Stable** (no move scoped): 696
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 837
-- **Realized polylogue/**/*.py**: 837 files declared
+- **Total declared**: 836
+- **Realized polylogue/**/*.py**: 836 files declared
 <!-- END GENERATED: topology-status -->
