--- polylogue/cli/action_contracts.py
@@ -1,35 +1,0 @@
-"""CLI compatibility imports for public action contracts."""
-
-from __future__ import annotations
-
-from polylogue.operations.action_contracts import (
-    ACTION_CONTRACT_BY_PATH,
-    ACTION_CONTRACTS,
-    PUBLIC_ACTION_FLOOR,
-    VIRTUAL_ACTION_PATHS,
-    ActionCardinality,
-    ActionEffect,
-    ActionFormat,
-    CliActionContract,
-    CompletionContext,
-    InputUnit,
-    MachineEnvelope,
-    action_completion_contexts,
-    contract_for_path,
-)
-
-__all__ = [
-    "ACTION_CONTRACTS",
-    "ACTION_CONTRACT_BY_PATH",
-    "PUBLIC_ACTION_FLOOR",
-    "VIRTUAL_ACTION_PATHS",
-    "ActionCardinality",
-    "ActionEffect",
-    "ActionFormat",
-    "CliActionContract",
-    "CompletionContext",
-    "InputUnit",
-    "MachineEnvelope",
-    "action_completion_contexts",
-    "contract_for_path",
-]
